Add files via upload
abnormal user behaviors on e-commerce platforms can be detected in various ways. for example, if an item in a shop is a best seller but many of its orders come from the same buyer, we have reason to suspect that seller is conducting order brushing. Order Brushing is a technique that may be employed by sellers to create fake orders in order to inflate the sellers or a particular items up the search result on shopee